app/requirements.txt
fastapi
uvicorn[standard]
boto3
botocore
python-dotenv
pydantic-settings
jinja2
Pillow